HDDS-9922. Migrate TestOzoneFileInterfaces to JUnit5
